# docs/mobile-app.openapi.yaml
openapi: 3.1.0
info:
  title: Waste Management System Mobile API
  version: 1.0.0
  summary: Mobile surveys and mobile tracking APIs for line supervisors
  description: |
    Unified OpenAPI contract for the mobile app.
    It covers:
    - line supervisor mobile login
    - line supervisor profile and assignment details
    - survey download and submission
    - mobile tracking status, ingestion, and deactivation
    - tracking admin monitoring endpoints
servers:
  - url: /
...
paths:
...
  /api/tracking/mobile/activate:
    post:
      tags: [Mobile Tracking]
      summary: Legacy device activation
      description: |
        Legacy compatibility endpoint.

        New mobile clients can login and send GPS batches directly to `/api/tracking/mobile/ingest`
        using the mobile bearer token without calling this endpoint first.
      operationId: activateMobileTracking
      deprecated: true
      security:
        - bearerAuth: []
        - sessionCookieAuth: []
        - secureSessionCookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileActivateRequest'
      responses:
        '201':
          description: Device activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileActivateResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /api/tracking/mobile/ingest:
    post:
      tags: [Mobile Tracking]
      summary: Ingest a GPS batch from the mobile application
      description: |
        Standard flow:
        1. Login using `/api/tracking/mobile/login`
        2. Send GPS batches directly to this endpoint with `Authorization: Bearer <accessToken>`

        `X-Tracking-Token` is still accepted for legacy clients but is no longer required for new ones.
      operationId: ingestMobileTrackingBatch
      security:
        - bearerAuth: []
        - sessionCookieAuth: []
        - secureSessionCookieAuth: []
        - trackingTokenHeader: []
      parameters:
        - in: header
          name: X-Tracking-Token
          required: false
          schema:
            type: string
          description: Legacy tracking token returned by `/api/tracking/mobile/activate`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTrackingBatchRequest'
            example:
              deviceId: android-device-001
              deviceName: Samsung A55
              platform: android
              appVersion: 1.0.0
              batchId: batch-20260330-001
              sentAt: '2026-03-30T10:00:30Z'
              samples:
                - recordedAt: '2026-03-30T10:00:05Z'
                  lat: 33.5138
                  lng: 36.2765
                  speed: 24
                  heading: 140
                  accuracy: 8
                  altitude: 650
      responses:
        '200':
          description: Batch stored and processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTrackingBatchResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: MobileAuthToken
    trackingTokenHeader:
      type: apiKey
      in: header
      name: X-Tracking-Token
      description: Legacy tracking token returned by `/api/tracking/mobile/activate`
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
    secureSessionCookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: الطلب غير صالح
            code: BAD_REQUEST
            status: 400
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: يجب تسجيل الدخول أولاً
            code: MOBILE_AUTH_REQUIRED
            status: 401
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: هذه الواجهة متاحة لمشرف الخط فقط
            code: MOBILE_ROLE_NOT_ALLOWED
            status: 403
...
    Error500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: حدث خطأ غير متوقع في الخادم
            code: INTERNAL_SERVER_ERROR
            status: 500
  schemas:
    ErrorResponse:
      type: object
      required: [message, code, status]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
...
    TrackingProvider:
      type: string
      enum: [athar, mobile_app, traccar]
    ZoneEventProvider:
      type: string
      enum: [athar, mobile_app]
...
    MobileActivateRequest:
      type: object
      properties:
        deviceId: { type: string }
        deviceName: { type: string }
        platform: { type: string }
        appVersion: { type: string }
    MobileTrackingBinding:
      type: object
      additionalProperties: true
    MobileActivateVehicle:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        plateNumber:
          type: [string, 'null']
        trackingProvider:
          $ref: '#/components/schemas/TrackingProvider'
        acceptedTrackingProviders:
          type: array
          items:
            $ref: '#/components/schemas/TrackingProvider'
        zoneEventProvider:
          $ref: '#/components/schemas/ZoneEventProvider'
    MobileActivateResponse:
      type: object
      properties:
        trackingToken: { type: string }
        binding:
          $ref: '#/components/schemas/MobileTrackingBinding'
        vehicle:
          $ref: '#/components/schemas/MobileActivateVehicle'
...
    MobileTrackingBatchRequest:
      type: object
      required: [batchId, samples]
      properties:
        deviceId:
          type: [string, 'null']
        deviceName:
          type: [string, 'null']
        platform:
          type: [string, 'null']
        appVersion:
          type: [string, 'null']
        batchId:
          type: string
        sentAt:
          type: [string, 'null']
          format: date-time
        samples:
          type: array
          items:
            type: object
            required: [recordedAt, lat, lng]
            properties:
              recordedAt:
                type: string
                format: date-time
              lat:
                type: number
              lng:
                type: number
              speed:
                type: [number, 'null']
              heading:
                type: [number, 'null']
              accuracy:
                type: [number, 'null']
              altitude:
                type: [number, 'null']
    MobileTrackingBatchResponse:
      type: object
      properties:
        success:
          type: boolean
        duplicate:
          type: boolean
        ingressId:
          type: [string, 'null']
        acceptedSamples:
          type: integer
        lateSamples:
          type: integer
        zoneEventsEnabled:
          type: boolean
        definitionsCount:
          type: integer
        legacyFallbackUsed:
          type: boolean
...